Renamed Web Api controllers to reflect the resource it exposes. Moved things around a bit, renamings, specifically the api controllers.
