+++ PDiffy/Features/Shared/Libraries/ImageDiffTool.cs
@@ -2,7 +2,7 @@ using System;
 using System.Drawing;
 using System.Drawing.Imaging;
 
-namespace PDiffy.Features.Shared
+namespace PDiffy.Features.Shared.Libraries
 {
 	public class ImageDiffTool : IImageDiffTool
 	{
